% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONV_GWH_EJ}
\alias{CONV_GWH_EJ}
\title{Conversion factor from gigawatt-hours to exajoules}
\format{
A numeric value.
}
\usage{
CONV_GWH_EJ
}
\description{
A numeric conversion factor used to convert energy values from GWh to EJ.
}
\keyword{datasets}
