% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EF.JEDI.RAW}
\alias{EF.JEDI.RAW}
\title{Raw JEDI employment factors}
\format{
A data frame.
}
\usage{
EF.JEDI.RAW
}
\description{
Raw employment factor data used in GCAM-USA jobs calculations.
}
\keyword{datasets}
