% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EF.JEDI}
\alias{EF.JEDI}
\title{Processed JEDI employment factors}
\format{
A data frame.
}
\usage{
EF.JEDI
}
\description{
Processed employment factor data used in GCAM-USA jobs calculations.
}
\keyword{datasets}
