% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCAM_GW.R
\name{GCAM_GW}
\alias{GCAM_GW}
\title{GCAM_GW}
\usage{
GCAM_GW(EJ_activity)
}
\arguments{
\item{EJ_activity}{Output from GCAM_EJ()}
}
\value{
A data frame with state-level capacity activity
}
\description{
GCAM_GW
}
\examples{
\dontrun{
GW_output <- GCAM_GW(EJ_activity)
}
}
