% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCAM_JOB.R
\name{GCAM_JOB}
\alias{GCAM_JOB}
\title{GCAM_JOB}
\usage{
GCAM_JOB(GW_output, method = NULL)
}
\arguments{
\item{GW_output}{Output from GCAM_GW()}

\item{method}{Specify method for construction job.
"Net": in a given year, a technology pre-mature retirement and newly installed net out.
"Total": in a given year, a technology pre-mature retirement and newly installed can happen at the same time.}
}
\value{
A data frame of all power sector direct job by subsector
}
\description{
GCAM_JOB
}
\examples{
\dontrun{
JOB_activity <- GCAM_JOB(GW_output, "Net")
JOB_activity <- GCAM_JOB(GW_output)
}
}
