% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCAM_model_step}
\alias{GCAM_model_step}
\title{GCAM model time step}
\format{
A numeric value.
}
\usage{
GCAM_model_step
}
\description{
Time step used between GCAM model years.
}
\keyword{datasets}
