% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcamusa.STATES}
\alias{gcamusa.STATES}
\title{GCAM-USA states}
\format{
A character vector or data frame.
}
\usage{
gcamusa.STATES
}
\description{
State-level identifiers used by GCAM-USA jobs calculations.
}
\keyword{datasets}
