% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_cf}
\alias{geo_cf}
\title{Geographic capacity factor data}
\format{
A data frame.
}
\usage{
geo_cf
}
\description{
Geographic capacity factor data used in GCAM-USA jobs calculations.
}
\keyword{datasets}
