% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OM_cost}
\alias{OM_cost}
\title{OM_cost}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8294 rows and 7 columns.
}
\source{
GCAM data, GCAM-USA 8.8; see inst/extdata/GCAM/GCAM8p8/Assumption.R
}
\usage{
OM_cost
}
\description{
This data technology level OM fixed and variable cost information
}
\details{
Generated by running \code{inst/extdata/GCAM/GCAM8p8/Assumption.R}.
}
\examples{
data(OM_cost)
}
\keyword{datasets}
