% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamdata.R
\name{read.GCAM.csv}
\alias{read.GCAM.csv}
\title{read.GCAM.csv}
\usage{
read.GCAM.csv(basename, na.strings = "")
}
\arguments{
\item{basename}{name of the file to be read in, without .csv}

\item{na.strings}{set to ""}
}
\value{
A data frame from raw csv file saved in the inst/extdata/GCAM folder
}
\description{
read raw csv files from inst/extdata/GCAM folder
}
\examples{
\dontrun{
A23.globaltech_retirement <- read.GCAM.csv("A23.globaltech_retirement")
}
}
