% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGCAM.R
\name{ReadGCAM}
\alias{ReadGCAM}
\title{ReadGCAM}
\usage{
ReadGCAM(
  filetype,
  input_path,
  db_name,
  query = system.file("extdata", "my_batch.xml", package = "GCAMUSAJobs"),
  scen_name,
  prj_name
)
}
\arguments{
\item{filetype}{specify the input is a database (folder) or a project (.dat)}

\item{input_path}{specify the full path for the input file}

\item{db_name}{specify the database name}

\item{query}{use the default query file in the package "my_batch.xml"}

\item{scen_name}{specify the scenario name (this needs to match the scenario name specified in the DB)}

\item{prj_name}{specify the project (.dat) name either to name it or load the existed one}
}
\value{
a list to be called in other functions
}
\description{
ReadGCAM
}
\examples{
\dontrun{
prj <- ReadGCAM(filetype = 'db',
input_path = "C:/GODEEP/GCAM7/output",
db_name = "database_basexdbGCAM-USA_Ref",
query = system.file("extdata", "my_batch.xml", package = "GCAMUSAJobs"),
scen_name = "GCAM-USA_Ref",
prj_name = "mydata.dat")
}
}
