% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCAM_futr_end}
\alias{GCAM_futr_end}
\title{GCAM future end year}
\format{
A numeric value.
}
\usage{
GCAM_futr_end
}
\description{
Ending year used for future GCAM model periods.
}
\keyword{datasets}
