% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCAM_futr_start}
\alias{GCAM_futr_start}
\title{GCAM future start year}
\format{
A numeric value.
}
\usage{
GCAM_futr_start
}
\description{
Starting year used for future GCAM model periods.
}
\keyword{datasets}
